// Client/app/main/InterFace/ILedger.ts
import * as common from './ICommon';
export class ILedger {
    ParentLedgerSysID: number;
    ParentLedgerName: string;
    LedgerSysID: number;
    LedgerName: string;
    AccountSysID: string;
    TaxSysID: string;
    TaxPercentage: number;
    IsMainLedger: boolean;
    IsAllowSubLedger: boolean;
    IsAutoGenerated: boolean;
    IsGlobal: boolean;
    ReferenceType: string;
    ReferenceSysID: number;
    BranchSysID: number;
    Credit: string;
    Debit: string;
    constructor() {
        this.LedgerSysID = 0;
    }
}